+++ rust/lance-datafusion/Cargo.toml
@@ -12,7 +12,9 @@ description = "Internal utilities used by other lance modules to simplify workin
 [dependencies]
 arrow.workspace = true
 arrow-array.workspace = true
+arrow-buffer.workspace = true
 arrow-schema.workspace = true
+arrow-select.workspace = true
 arrow-ord.workspace = true
 async-trait.workspace = true
 datafusion.workspace = true
