+++ rust/lance-datafusion/src/lib.rs
@@ -5,4 +5,7 @@ pub mod chunker;
 pub mod dataframe;
 pub mod exec;
 pub mod expr;
+pub mod logical_expr;
+pub mod planner;
+pub mod sql;
 pub mod utils;
