+++ rust/lance/src/datafusion.rs
@@ -5,5 +5,4 @@
 //!
 
 pub(crate) mod dataframe;
-pub(crate) mod logical_expr;
 pub(crate) mod logical_plan;
