+++ rust/lance/src/io/exec.rs
@@ -7,7 +7,6 @@
 
 pub(crate) mod knn;
 mod optimizer;
-mod planner;
 mod projection;
 mod pushdown_scan;
 pub mod scalar_index;
@@ -18,7 +17,9 @@ pub mod testing;
 pub mod utils;
 
 pub use knn::{ANNIvfPartitionExec, ANNIvfSubIndexExec, KNNVectorDistanceExec, PreFilterSource};
-pub use planner::{FilterPlan, Planner};
+pub use lance_datafusion::planner::Planner;
+pub use lance_index::scalar::expression::FilterPlan;
+pub use optimizer::get_physical_optimizer;
 pub use projection::ProjectionExec;
 pub use pushdown_scan::{LancePushdownScanExec, ScanConfig};
 pub use scan::LanceScanExec;
